docker-compose.yaml is no longer required in this project since I now have the one docker-compose file to rule them all.
Also, when I moved to a mega docker-compose file, I broke the auto-upgrade process as the docker_compose_location in secrets.yaml had changed, and I now need to specify the "-p home-server" argument so there isn't a conflict in container name.
It feels a little dirty as this couples this file to my other docker-compose file, but I can't be bothered making it any nicer since it's unlikely to change. :)