Adding in an automation to turn off the garden lights if they have been left on during the day, to save power. Also updating the automation which triggers when opening the alfresco doors to not turn on the flood lights, since this isn't really necessary.
In the process I discovered that the Fibaro double switch used for the two outdoor lights doesn't work reliably if I send a command to turn on/off both lights in quick succession. As a result I've had to update any scripts which look to turn on/off both of these lights to trigger twice with a slight delay between them, just in case it doesn't work the first time around. :/