Updating the "turn_off_all_lights" script to take in a list of lights to exclude when turning off all the lights in the house. This solves a number of problems:
- The garage light can remain on when setting the alarm, to help navigate in the dark garage (this will automatically turn off when the garage door is closed again anyway)
- The front porch light can remain on when leaving the house, in case we are coming home late at night
- The master bedroom lamp can be left on when pressing the "turn off house" button when going to bed at night
- The theatre room lamps don't need to turn off and then on again when setting movie lighting, but instead can just change straight to their target state

It looks like not providing a value for the "excluded_lights" field will still allow the data_template to work fine - it just doesn't run the rejectattr() step of the template. However this appears to be undocumented behaviour, so just to be safe I have included that field everywhere and intend on passing an empty array if I don't want to exclude any lights.

As part of this change I have also updated the master bedroom "turn off the house" automation to first turn off the brightest lights in the master bedroom (as these have the biggest immediate impact), then turn on the master bedroom ceiling lights to help with navigation to bed, then only to turn off the rest of the house (including the master bedroom ceiling) after an appropriate amount of time. This fixes the issue where the master bedroom ceiiling lights wouldn't turn on for about 5-10 seconds after pressing the button due to having to wait for everything else to turn off, by which point I was often already in bed.