Reworked the youtube cast integration usage to first lookup all songs in the playlist and randomize them, before adding them one-by-one to the cast playlist.
Also updated the Christmas and board games scripts to use this new approach from the youtube.yaml package.

Really, it's kind of a horrible approach to this; I had to constantly work around the limitations of the RESTFUL sensor, the jinja templating syntax, and the way that Home Assistant variables in scripts can't retain their value between loops in a repeat block. But it works in the end.

This approach also has a few limitations:
- The maximum number of songs which can be added from a playlist is 50
- The whole integration (both before and now) only allows for casting to devices with screens, such as the Google Home Hub, as it is playing from Youtube not Youtube Music
- Currently polling of the Youtube API has to be at <3 seconds (set to 1 second currently) as reloading the sensor doesn't seem to force it to refresh the value

A better approach which I'll probably migrate to later is to replace the RESTFUL sensor with a custom integration which does the same thing, but randomizes the results before they come back. This would have the following advantages:
- Removes the need for the ugly jinja templating
- Can iterate over the Youtube API pages to allow for more than 50 songs to be retrieved
- Don't need to poll the API as frequently (maybe), can just call it on demand as needed, and perhaps once daily if a poll time is mandated